+++ R/crgt.R
@@ -1,0 +1,30 @@
+# Copyright 2023 Craig P. Hutton
+#
+# Licensed under the Apache License, Version 2.0 (the “License”); you may not
+# use this file except in compliance with the License. You may obtain a copy of
+# the License at:
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an “AS IS” BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations under
+# the License.
+
+#' "crgt: data processing and analysis for the cued Rat Gambling Task."
+#'
+#' "This package provides functions for processing and analysis of operant
+#' behavioural data for the cued and uncued versions of the Rat Gambling Task
+#' (RGT) and the cRGT training version of the five choice serial reaction time
+#' task (5CSRT). Provided functions enable the user to parse raw MedPC data
+#' files, pivot them between long and wide formats, enabling them to perform
+#' common statistical analyses for these behavioural tasks as used in the
+#' Winstanley laboratory at the University of British Columbia. This package
+#' aims to facilitate reproducible and standardized analysis of RGT data. Note
+#' This package is currently in development and its functionality will be
+#' improved over time."
+#'
+#' @docType package
+#' @name crgt
+"_PACKAGE"